Closes #105 - heredoc now doesn't segfault on Ctrl+D but prints a warning message and lets the rest of the code run.
